+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**Internal**
+
+- Postpone metrics aggregation until we received the project state. ([#2588](https://github.com/getsentry/relay/pull/2588))
+- Remove `AggregationService` and rename `RouterService` to `AggregationService`. ([#2611](https://github.com/getsentry/relay/pull/2611))
+
 ## 23.11.0
 
 **Features**:
@@ -42,9 +47,7 @@
 - Introduce a dedicated usage metric for transactions that replaces the duration metric. ([#2571](https://github.com/getsentry/relay/pull/2571), [#2589](https://github.com/getsentry/relay/pull/2589))
 - Restore the profiling killswitch. ([#2573](https://github.com/getsentry/relay/pull/2573))
 - Add `scraping_attempts` field to the event schema. ([#2575](https://github.com/getsentry/relay/pull/2575))
-- Postpone metrics aggregation until we received the project state. ([#2588](https://github.com/getsentry/relay/pull/2588))
 - Move `condition.rs` from `relay-sampling` to `relay-protocol`. ([#2608](https://github.com/getsentry/relay/pull/2608))
-- Remove `AggregationService` and rename `RouterService` to `AggregationService`. ([#2611](https://github.com/getsentry/relay/pull/2611))
 
 ## 23.9.1