+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Remove from events spans starting or ending before January 1, 1970 UTC. ([#2627](https://github.com/getsentry/relay/pull/2627))
+- Remove from events breadcrumbs dating before January 1, 1970 UTC. ([#2635](https://github.com/getsentry/relay/pull/2635))
 
 **Internal**: